move tests to unarchive_test to unarchive package
